backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.75.0",
    "fastapi[standard]>=0.122.0",
    "google-genai>=1.52.0",
    "google-generativeai>=0.8.5",
    "pillow>=12.0.0",
]
